Created a standard SVN folder structure and moved the Sprite Viewer into the trunk folder.
trunk - The main source directory.  This is the primary commit folder for code.
branches - Used whenever a branch is necessary.  To branch, copy the trunk into the branch directory and commit to the branch rather than the trunk.
docs - Used for non-code files (aka documents, aka class diagrams etc.)
tags - Used for actual releases of the product line.  Whenever a new version is "released" copy the trunk to this directory with a tagname.