Updated the display module to match the new class diagram and began implementation.
Most of our work was in the Display class, where we created the main window and implemented how layers are to be drawn to the screen.

Added GameStateManager which is responsible for handling the references to the GameState objects.  It determines which are the display, logic, and buffer game states.

Added GameState as a filler class with a single method for getting the size of the map.

That is all.

-Titus Klinge
-Ryan Tew