Major updates.
Implemented MapItemDrawer
Implemented Animation

Minor changes to Display so that the positions are represented as doubles instead of ints

Deleted unused classes

Implemented most of CommandCardPanel.  The buttons display correctly, however, because the input part of the module is not defined, the updateButtons() method is not implemented (it is in sudo code though).

Added some placer methods in the GameState class (aka getTime()).

That's about it folks.  Oh yeah, deleted the bin directory.  (Connor.....why, oh why...?)

-Titus Klinge
-Ryan Tew