A lot of work.  Sorry for the large commit.
Worked a lot on the panels of the game.  We implemented each panel to a certain extent, but were not able to put information in that atm.  However, they show up and look cool.

Deleted a bunch of unused classes such as Button, TooltipPanel, DrawableMap, Drawable, DrawableUnit, and DrawableBuilding.

Moved the panels into their own package.

Updated the display to correctly draw the panels layer.

Did a little work on the graph layer (aka the edge and node layer for strategic view).  Mostly sudo code though until we have a more concrete specification for dependant classes.

Cool beans.

-Titus Klinge
-Ryan Tew