Completed 4 abilities: -Construct Building - builds a building -Construct Unit - builds a unit -Research Tech - researchs some tech -Shoot
construct building and research tech are auto-generated from the list of the race's buildings and tech and added as abilities to the command center. Only the command center may have the ability construct building. Only buildings may have the ability construct unit. Only a unit may have the ability shoot. (this is all as of this moment, but it can always be changed in the future fairly simply). Only combat strategies should create shoot; only orders should create construct building and research tech; only w/e constructs a building (in this case the construct building ability) can create construct unit abilities. 

As of right now, none of this code is tested, so it may have logical bugs. But right now Eclipse says there are no compilation issues.

Finally, I added a function class. It simulates a mathematical function. It has an enum to determine what type of function it is (atm just polynomial and exponential) and its constructor takes in a parser.

-Connor Schenck