I added everything that I think we are going to need to draw colored bezier curves on top of the lanes in strategic view. This included: 	a normalize method in the Position class (now that I think about it I should have added a getOrthagonal method too) 	4 positions in the Lane class to represent the curve, along with their getters 	a width in the Lane class, along with its getter 	making a call to setOpaque(false) in the Display 	adding the GraphLayer to the Strategic view in the Display 	I added some test code to the GraphLayer just to make sure the ColoredEdge class was drawing something reasonable
The implementation for drawing bezier curves is not complete. The gist of it is complete and it draws the general shape, but it leaves gaps on the outside of the curve. I plan on fixing that by creating the rectangle's ends to cut across the lane instead of just being a rectangle with a width the same as the lane. I have not thought too much about how to do this though. Another thing that needs addressed with this is how many points are used to represent the curve. Right now the algorithm is using 11 points, but the wiki page suggests using more points in curvier sections and less points in more linear sections by using recursive subdivision.

-Ryan Tew