Units now are spawned in the correct place and now move as they should.
The problem with them being spawned in the wrong spot was due to BezierCurve code computed degrees while code that used it expected radians.  This is something that should've been specified during design IMO that I never thought of.

The problem with units moving improperly was a misuse of the Shape.transform() method.  Units were translating BY their position instead of TO their position.  If you're using a method, READ ITS DOCUMENTATION INSTEAD OF MAKING ASSUMPTIONS ABOUT HOW IT WORKS.

:D

Taylor