Made some changes while searching for the desync issue.  Turns out there was a System.currentTimeMillis() call in the pathfinding code.
I implemented equals() on a lot of things while searching; now, LogicBlockingManager checks if its two GameStates are equal whenever they have been updated to the same tick.  If they aren't, it will say that it detected desync and give the tick number.

I also made a small change to BezierCurve when I thought I had found the desyncing problem; that wasn't it the problem, but the change is positive if small, so I left it.

-Taylor