Added sounds to the Mech race!!! Some sounds are not used at all, as I wasn't sure where they were supposed to be used. Some things that probably should have sounds don't, as I wasn't sure which sound was supposed to be used.
Fixed a bug in the MapItemCommanalitiesEditor where it would crash if the lastDirectory.txt file was not present.

Changed how the random number generator in the AnitProjectileConfiguration targeting strategy was seeded to get more random-seeming behavior.

Things that (IMO) need to be improved about the sound:
-Fix the latency issue!! (VERY big deal, but possibly not trivial)
-Every game, I heard the tank tread sound when I had built no tanks.
Possibly a race configuration issue?
-One game, the sound stopped playing after a while (might have been prompted by alt-tabbing)
-Sound effects are far quieter than the music by default; it's impossible to hear them.
-Many events in the game (like units of the same type from the same wave shooting) occur at precisely the same instant,
and cannot aurally be distinguished from a single event. Perhaps add in some small random noise to when these occur?
-Some sounds are too loud, while others are too quiet.
In many cases, this can be resolved by adjusting the volume of the file,
but in some cases sounds that should be played at different volumes
are loaded from the same file (Gates exploding vs. units exploding)